Replacing these param lists with ImmutableList
Mainly so that they look different to IDEA. Multiple sequential statements show up as duplicated code. Two immutable lists with different contents don't. There's probably a way to share the common part of the lists still. Perhaps a private method could create a builder which builds the list with the first three parameters, and then the method appends the last two and builds the list?